Replaced old page with new page
