+++ client/src/components/Navbar.js
@@ -35,11 +35,6 @@ export default function Navbar() {
       id: 7,
       link: "#",
     },
-    {
-      name: 'TEST',
-      id: 8,
-      link: '/TEST'
-    },    
   ];
   const log = [
     {
@@ -67,16 +62,6 @@ export default function Navbar() {
       id: 5,
       link: "#",
     },
-    {
-      name: 'TEST',
-      id: 8,
-      link: '/TEST'
-    },
-    {
-      name: 'newcampaign',
-      id: 9,
-      link: '/newcampaign'
-    }
   ];
 
   const [anchorElNav, setAnchorElNav] = React.useState(null);
